Refactor: Move implementation of StakerspeculatorAgent into SpeculatorAgent, and reuse code from it. DRY.
