+++ agents/SpeculatorAgent.py
@@ -1,3 +1,10 @@
+"""Implement SpeculatorAgent *and* StakerspeculatorAgent.py
+-parent: SpeculatorAgentBase
+-child: SpeculatorAgent - speculate via buying & selling DT
+-child: StakerspeculatorAgent - speculate via staking & unstaking
+"""
+
+from abc import abstractmethod
 from enforce_typing import enforce_types
 import random
 from typing import List
@@ -10,9 +17,8 @@ from util import constants
 DEFAULT_s_between_speculates = 1 * constants.S_PER_DAY
 
 @enforce_types
-class SpeculatorAgent(AgentBase.AgentBaseEvm):
-    """Speculates by buying and selling DT"""
-
+class SpeculatorAgentBase(AgentBase.AgentBaseEvm):
+    
     def __init__(self, name: str, USD: float, OCEAN: float,
                  s_between_speculates:int = DEFAULT_s_between_speculates):
         super().__init__(name, USD, OCEAN)
@@ -33,6 +39,23 @@ class SpeculatorAgent(AgentBase.AgentBaseEvm):
             return False
         else:
             return self._s_since_speculate >= self._s_between_speculates
+        
+    @abstractmethod
+    def _speculateAction(self, state):
+        pass
+
+    def _poolsForSpeculate(self, state) -> List[AgentBase.AgentBaseAbstract]:
+        pool_agents = state.agents.filterToPool()
+
+        if hasattr(state, 'rugged_pools'):
+            for pool_name in state.rugged_pools:
+                del pool_agents[pool_name]
+
+        pool_agents = pool_agents.values()
+        return pool_agents
+
+class SpeculatorAgent(SpeculatorAgentBase):
+    """Speculates by buying and selling DT"""
 
     def _speculateAction(self, state):
         pool_agents = self._poolsForSpeculate(state)
@@ -53,12 +76,21 @@ class SpeculatorAgent(AgentBase.AgentBaseEvm):
             DT_buy_amt = 1.0  # magic number
             self._wallet.buyDT(pool, datatoken, DT_buy_amt, max_OCEAN_allow)
 
-    def _poolsForSpeculate(self, state) -> List[AgentBase.AgentBaseAbstract]:
-        pool_agents = state.agents.filterToPool()
+@enforce_types
+class StakerspeculatorAgent(SpeculatorAgentBase):
+    """Speculates by staking and unstaking"""
 
-        if hasattr(state, 'rugged_pools'):
-            for pool_name in state.rugged_pools:
-                del pool_agents[pool_name]
+    def _speculateAction(self, state):
+        pool_agents = self._poolsForSpeculate(state)
+        assert pool_agents, "need pools to be able to speculate"
 
-        pool_agents = pool_agents.values()
-        return pool_agents
+        pool = random.choice(list(pool_agents)).pool
+        BPT = self.BPT(pool)
+
+        if BPT > 0.0 and random.random() < 0.50:  # magic number
+            BPT_sell = 0.10 * BPT  # magic number
+            self.unstakeOCEAN(BPT_sell, pool)
+
+        else:
+            OCEAN_stake = 0.10 * self.OCEAN()  # magic number
+            self.stakeOCEAN(OCEAN_stake, pool)
