--- agents/test/test_StakerspeculatorAgent.py
@@ -1,2 +1,0 @@
-#all tests for StakerspeculatorAgent are in test_SpeculatorAgent.py
-# in a pytest-parameterized fashion. Why: DRY
